Final commit

+++ booking-impl/src/main/java/za/co/interview/study/impl/dto/BookingSlotDto.java
@@ -15,6 +15,6 @@ public class BookingSlotDto {
 	Date startTime;
 	Date endTime;
 	int duration;
-	int roomId;
+	long roomId;
 	boolean maintenanceSlot;
 }
